연월별(시험별) 학생들의 과목 성적 조회 API 추가 구현

+++ src/main/java/com/project/lms/vo/ScoreAvgStatsListByClassVO.java
@@ -11,8 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ScoreAvgStatsListByClassVO{
     @Schema(description = "과목별 평균 점수 리스트")
-    // private List<ScoreAvgListBySubjectVO> scoreList = new ArrayList<>();
-    private Map<String, Integer> scoreList;
+    private Map<String, Double> scoreList;
     @Schema(description = "반 이름")
     private String className;
